added empty methods and getters and setters to all impl classes

+++ src/facility_maintenance/ScheduleMaintenance.java
@@ -5,7 +5,7 @@ public interface ScheduleMaintenance {
     public String getMaintDate();
     public void setMaintDate(String maintDate);
     public void scheduleMaintenance();
-    public Double calcDownTimeForFacility();
+    public int calcDownTimeForFacility();
     public Double calcMaintenanceCostForFacility();
 
 }
